Refactoring into new package
